+++ src/components/Docs/MetricVisualizations/index.tsx
@@ -1,0 +1,4 @@
+export { default as AreaEnclosedVisualization } from '@/components/Docs/MetricVisualizations/AreaEnclosedVisualization';
+export { default as BorderTracedVisualization } from '@/components/Docs/MetricVisualizations/BorderTracedVisualization';
+export { default as PathPrecisionVisualization } from '@/components/Docs/MetricVisualizations/PathPrecisionVisualization';
+export { default as RouteEfficiencyVisualization } from '@/components/Docs/MetricVisualizations/RouteEfficiencyVisualization';